refactor: split app builder into separate file
